CruxTaskBot/requirements.txt
discord.py
python-dotenv
jishaku
colorama
sqlmodel